requirements.txt
adb==1.3.0
aiohttp==3.7.4.post0
aiohttp-sse-client==0.1.7
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
configparser==5.0.1
crc16==0.1.1
cryptography==3.4.6
Django==2.2.19
django-cors-headers==3.7.0
django-environ==0.4.5
django-model-utils==4.1.1
django-polaris==1.3.2
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libusb1==1.9.1
mccabe==0.6.1
mnemonic==0.19
multidict==5.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stellar-base-sseclient==0.0.21
stellar-sdk==2.11.3
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
